debug: upgrade to solidity-ibc-eureka latest commit
